Fixed mobile layout
Fixes all the bugs I can spot on mobile, though stops short of
implementing a hamburger menu. I think that will be necessary once I add
another item to the top menu besides the blog, but it's time for me to
go to bed so I'll leave that for another day.

Also includes some prettier fixes