Add About Me page
Adds a scrollytelling about me page that will live at the base route on
the site. Hopefully all the interactions are intuitive and the content
gives a good introduction to who I am!

Implementing this involved several changes:

- Creating a `Scroller` component (and a few other components to go
along with it) that wraps around Scrollama and provides an easy to use
interface for scrollytelling
- Updating the theme to have a lighter blue as the secondary color.
This not only looks better for the scroller cards, but it also makes
things like buttons pop more.
- Improved the way images are loaded in by adding the
`svelte-preprocess-import-assets` plugin. Means I don't need to add
`import`s for every image