Projects accordion
On mobile, the table layout for projects was getting pretty
squished and a lot of info was hidden entirely (e.g. the description
and the `builtWith`/`deployedWith` lists).

So instead, on small screens I'm now showing an accordion view instead
of a table. This lets users click into specific projects to get more
info about them.